backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
boto3
langchain
openai
pypdf
python-docx
docx2txt
elasticsearch
pydantic_settings
